+++ CHANGELOG.md
@@ -13,6 +13,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - **dropdown**: add no-arrow attribute to hide the arrow in dropdown button
 - **angular/formly**: add disabled attribute support for field template options
 - **switch**: add disabled attribute
+- **input**: add `prefixLabel` and `suffixLabel` attributes
 
 ### Changed